Feat: 카프카 로직 제거하고 사용자 로그인 기록 저장을 처리해라 - #129
