+++ tictoc-api/src/main/java/tictoc/user/port/UserCommandUseCase.java
@@ -3,5 +3,5 @@ package tictoc.user.port;
 import tictoc.config.security.jwt.dto.JwtResDTO;
 
 public interface UserCommandUseCase {
-    JwtResDTO.Login login(String authenticationCode);
+    JwtResDTO.Login login(String authenticationCode, String clientIp, String userAgent);
 }
