+++ tictoc-domain/src/main/java/tictoc/user/adapter/UserRepositoryAdapter.java
@@ -11,13 +11,12 @@ import static tictoc.error.ErrorCode.USER_NOT_FOUND;
 
 @Adapter
 @RequiredArgsConstructor
-public class UserRepositoryAdaptor implements UserRepositoryPort {
+public class UserRepositoryAdapter implements UserRepositoryPort {
     private final UserRepository userRepository;
 
     public User saveUser(User user) {
         return userRepository.save(user);
     }
-
     public Optional<User> findUserByKakaoId(String kakaoId) {
         return userRepository.findByKakaoId(kakaoId);
     }
@@ -25,7 +24,6 @@ public class UserRepositoryAdaptor implements UserRepositoryPort {
     public User findUserById(Long userId) {
         return userRepository.findById(userId).orElseThrow(() -> new UserNotFoundException(USER_NOT_FOUND));
     }
-
     public boolean existsUserByKakaoId(String kakaoId) {
         return userRepository.existsByKakaoId(kakaoId);
     }
