+++ tictoc-domain/src/main/java/tictoc/user/port/UserLoginHistoryRepositoryPort.java
@@ -1,0 +1,8 @@
+package tictoc.user.port;
+
+import tictoc.user.model.UserLoginHistory;
+import java.time.LocalDateTime;
+
+public interface UserLoginHistoryRepositoryPort {
+    UserLoginHistory save(Long userId, LocalDateTime loginAt, String userIp, String userAgent);
+}
