--- tictoc-infrastructure/src/main/java/tictoc/kafka/exception/KafkaPublishException.java
@@ -1,10 +1,0 @@
-package tictoc.kafka.exception;
-
-import tictoc.error.ErrorCode;
-import tictoc.error.exception.TicTocException;
-
-public class KafkaPublishException extends TicTocException {
-    public KafkaPublishException(final ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
