--- tictoc-infrastructure/src/main/java/tictoc/kafka/exception/KafkaSettingException.java
@@ -1,10 +1,0 @@
-package tictoc.kafka.exception;
-
-import tictoc.error.ErrorCode;
-import tictoc.error.exception.TicTocException;
-
-public class KafkaSettingException extends TicTocException {
-    public KafkaSettingException(final ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
